Move pattern ns to top level, add tests
In order for test.check to be an optional dependency we can't require the
generators ns at the top level. This means the top level really becomes just a
proxy for pattern, in which case we might as well promote pattern to the top.

Add some semblance of initial example based tests.