Use non-capturing groups for incidental grouping
This is in preparation of introducing explicit capturing
groups. However, it turns out that this is desirable on its own
because using capturing groups for incidental grouping purposes as
regal does right now has an effect on the result of `re-find` and
friends:

    (re-find #"a(b|c)" "ab")   => ["ab" "b"]
    (re-find #"a(?:b|c)" "ab") => "ab"

So users whould have had to figure out whether a regal expression
would compile to such incidental groups to reliably handle match
results.