Introduce `grouped?` multimethod for ops
Ops may now be marked as grouped on the form level via the `grouped?`
multimethod, rather than on the regex level via the `::grouped`
metadata annotation. This allows enclosing forms to determine whether
a child form is grouped without being forced to prematurely convert it
to regex (see `suffix-op->regex` for use case).

Also, suffix ops are not marked as grouped anymore as that would lead
to invalid regexen like `[:* [:* \x]] => #"x*x"`.