Make Regal-as-spec composable
Introduce `lambdaisland.regal.spec-alpha/spec`, which converts a Regal
expression into a spec which both validates and generates. When used as such any
qualified keywords will be resolved using the global spec registry, which
assumes that these are also Regal specs.

This adds an options map to both the pattern compilation and generator code
paths that gets threaded through, to provide the resolver to nested expressions.

This commit also renames regal->grouped to regal->ir, since we have a confusing
double meaning of "grouped" otherwise (since we also have an explicit ::grouped
metadata).

Also makes some things in generator private, and adds a bit of docstring.