Round-trip the existing test cases
Make sure that all cases in test_cases.edn parse back to the given form. This
also meant dealing with the "synthetic" implementations of :vertical-tab /
:vertical-whitespace, and implementing some missing features: :lookahead,
:lookbehind, :atomic, ...

Streamline the API so there are convenient conversions between all three
formats, there's an overview in the README.

Update the README somewhat, although the grammar description is still outdated.