Fix whitespace-related issues
The `(?s)` modifier (makes `.` match newlines) is not recognized in
JavaScript (at least not in all JS environments), replace it in the instaparse
grammar with something that works everywhere.

Make the `:whitespace` / `:non-whitespace` generators consistent with the
semantics of the corresponding regex patterns.

Make sure `[:cat :line-break :line-break]` can't generate `\r\n`, since this is
considered a single line break, not a double line break.