Anchors: emit ^ and $ on all platforms, add parsing
The Java docs are confusing on this part, but in the absence of modifier flags ^
and $ do in fact behave the same on Java and JavaScript, it is only when
multiline mode (?m) is enabled that \A and \z come into the picture.

We'll emit ^ and $ for :start and :end respectively, but parse \A and \z as well.