Parsing fixes
Parse unrecoginzed escape code to the character they escape, instead of
returning a parse error. This is generally how regex engines work, you are
allowed to add unnecesary backslashes.

On the java flavor: parse `\s` to

    [:class " " :tab :newline :vertical-tab :form-feed :return]

and `\S` to `[:not ...]`. The semantics we use for `:whitespace` do not match
those of `\s` on Java, we match a much broader set of unicode whitespace
characters.