Specify a garbage collector
I'm wondering if the G1GC garbage collector avoids too much garbage accumulation, which is why later Java versions avoid getting the kill signal as much.

This theory is imperfect, as Java 9 was the first to enable it, yet those jobs frequently failed. My explanation for that is that maybe some of our CI hardware gets the older GC on Java 9.