Add lazy quantifiers
The default behavior of a regex quantifier is "greedy": consume as many
characters as possible to satisfy the match.

    user> (re-find #".*!" "foo! bar!")
    "foo! bar!"

Sometimes it's useful to be able to grab the fewest characters
satisfying a quantifier. This is where the ["lazy" variants of
quantifiers][1] come in:

    user=> (re-find #".*?!" "foo! bar!")
    "foo!"

This patch adds lazy versions to `*`, `+`, and `?`:

    user=> (re-find (regal/regex [:cat [:*? :any] "!"]) "foo! bar!")
    "foo!"

[1]: http://www.rexegg.com/regex-quantifiers.html#lazy